refatorção da arquitetura de rotas e nova rota getBirthdates mobile
